ai_vision_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
python-decouple==3.8
google-generativeai
python-dotenv
